ai_photostudio/requirements.txt
fastapi
uvicorn
python-telegram-bot==20.7
replicate
python-dotenv
